requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
transformers==4.35.2
torch==2.1.1
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0